2022/11/29 FarmStory2 모델2 게시판 프로젝트 실습

+++ Jboard2/src/main/java/kr/co/jboard2/controller/DeleteController.java
@@ -45,7 +45,9 @@ public class DeleteController extends HttpServlet {
 		String newName = (String)map.get("newName");
 		
 		if(result > 0) { // 게시물 삭제 성공
-			service.deleteFile(req, "/file", newName);
+			if(newName != null) {
+				service.deleteFile(req, "/file", newName);
+			}
 			resp.sendRedirect(req.getContextPath() + "/list.do");
 			return;
